Move actions into new namespace - rename base action as a defined action
