+++ lib/cucumber/core/test/action.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require 'cucumber/core/test/location'
+require 'cucumber/core/test/result'
+require 'cucumber/core/test/timer'
+
+require 'cucumber/core/test/action/defined'
+require 'cucumber/core/test/action/undefined'
+require 'cucumber/core/test/action/unskippable'
