+++ lib/cucumber/core/test/action/unskippable.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+require 'cucumber/core/test/result'
+require 'cucumber/core/test/action/action'
+
+module Cucumber
+  module Core
+    module Test
+      module Action
+        class Unskippable < Action
+          def skip(*args)
+            execute(*args)
+          end
+        end
+      end
+    end
+  end
+end
