--- lib/cucumber/core/test/actions.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require 'cucumber/core/test/location'
-require 'cucumber/core/test/result'
-require 'cucumber/core/test/timer'
-
-require 'cucumber/core/test/actions/action'
-require 'cucumber/core/test/actions/undefined_action'
-require 'cucumber/core/test/actions/unskippable_action'
